Splitting and dynamic-loading inspect mode commands
